Use stack for correct kdtree traversal
When a ray intersects both children both must be traversed. It is easier
to reason about keeping the to-be-traversed children on a stack instead
of restarting at the root and relying on the t-range to get correct
traversal. It appears to also be faster depending on the scene's tree
depth, probably because we can avoid re-traversing the top of the tree
multiple times.